pyproject.toml
name = "kasparro-backend"
version = "1.0.0"
description = "Production-grade backend data ingestion system"
authors = [{name = "Sabiha Anjum"}]

[project]
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "alembic==1.13.0",
    "pydantic==2.5.0",
    "psycopg2-binary==2.9.9",
    "aiohttp==3.9.1",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.23.2",
    "pytest-cov==4.1.0",
    "httpx==0.25.2",
    "black==23.12.0",
    "isort==5.13.2",
    "flake8==6.1.0",
    "mypy==1.7.1",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
